renamed db script
