+++ gef-portal-scraper/gef_portal_scraper/document_db_manager.py
