+++ gef-portal-scraper/gef_portal_scraper/gef_portal_scraper.py
@@ -4,7 +4,7 @@ import logging
 import os
 import sys
 from urllib.parse import unquote, urlparse
-import create_db_tables
+import document_db_manager
 
 import requests
 from bs4 import BeautifulSoup
@@ -77,7 +77,7 @@ def download_file(href, project_id, idx):
         logging.info(f"Downloaded file: {file_path}")
         # store document info in db
         filename, extension = os.path.splitext(href)
-        create_db_tables.add_document(original_filename, href, extension, project_id)
+        document_db_manager.add_document(original_filename, href, extension, project_id)
         return file_path
     except requests.RequestException as e:
         logging.error(f"Failed to download file from: {href}. Error: {e}")
@@ -103,7 +103,7 @@ def download_files_from_project_page(project_id):
         if href:
             file_extension = os.path.splitext(href)[1]
             if file_extension in VALID_EXTENSIONS:
-                if create_db_tables.document_exists(href):
+                if document_db_manager.document_exists(href):
                     logging.info(f"Document already in database, skipping: {href}")
                     continue
                 downloaded_file = download_file(href, project_id, idx)
@@ -127,7 +127,7 @@ def download_project_ids(project_ids: list[str]):
 def main():
     setup_logging()
 
-    create_db_tables.create_db_and_tables()
+    document_db_manager.create_db_and_tables()
 
     create_directory(OUTPUT_PATH)
 
@@ -136,7 +136,7 @@ def main():
 
     # project_ids = get_ids_from_json("gef_7_project_ids")
 
-    download_project_ids(project_ids[0:5])
+    download_project_ids(project_ids[0:10])
 
 
 if __name__ == "__main__":
